Moved exceptions to api module
