+++ api/src/main/java/com/tinqinacademy/hotel/api/exception/BaseException.java
@@ -1,4 +1,4 @@
-package com.tinqinacademy.hotel.core.exception;
+package com.tinqinacademy.hotel.api.exception;
 
 import lombok.Getter;
 import org.springframework.http.HttpStatus;
