+++ api/src/main/java/com/tinqinacademy/hotel/api/exception/exceptions/PartialUpdateRoomException.java
@@ -1,6 +1,6 @@
-package com.tinqinacademy.hotel.core.exception.exceptions;
+package com.tinqinacademy.hotel.api.exception.exceptions;
 
-import com.tinqinacademy.hotel.core.exception.BaseException;
+import com.tinqinacademy.hotel.api.exception.BaseException;
 import org.springframework.http.HttpStatus;
 
 public class PartialUpdateRoomException extends BaseException {
