+++ api/src/main/java/com/tinqinacademy/hotel/api/exception/exceptions/ViolationException.java
@@ -1,9 +1,7 @@
-package com.tinqinacademy.hotel.core.exception.exceptions;
+package com.tinqinacademy.hotel.api.exception.exceptions;
 
 
-import com.tinqinacademy.hotel.core.exception.BaseException;
 import lombok.Getter;
-import org.springframework.http.HttpStatus;
 
 import java.util.List;
