+++ core/src/main/java/com/tinqinacademy/hotel/core/processors/BaseOperationProcessor.java
@@ -2,9 +2,8 @@ package com.tinqinacademy.hotel.core.processors;
 
 import com.tinqinacademy.hotel.api.base.OperationInput;
 import com.tinqinacademy.hotel.core.errors.ErrorMapper;
-import com.tinqinacademy.hotel.core.exception.exceptions.ViolationException;
+import com.tinqinacademy.hotel.api.exception.exceptions.ViolationException;
 import jakarta.validation.ConstraintViolation;
-import jakarta.validation.ConstraintViolationException;
 import jakarta.validation.Validator;
 import lombok.AllArgsConstructor;
 import org.springframework.core.convert.ConversionService;
