+++ core/src/main/java/com/tinqinacademy/hotel/core/processors/system/RegisterVisitorOperationProcessor.java
@@ -6,8 +6,8 @@ import com.tinqinacademy.hotel.api.operations.system.registervisitor.RegisterVis
 import com.tinqinacademy.hotel.api.operations.system.registervisitor.RegisterVisitorOutput;
 import com.tinqinacademy.hotel.api.operations.system.registervisitor.RegisterVisitorOperation;
 import com.tinqinacademy.hotel.core.errors.ErrorMapper;
-import com.tinqinacademy.hotel.core.exception.exceptions.NotFoundException;
-import com.tinqinacademy.hotel.core.exception.exceptions.RegisterVisitorException;
+import com.tinqinacademy.hotel.api.exception.exceptions.NotFoundException;
+import com.tinqinacademy.hotel.api.exception.exceptions.RegisterVisitorException;
 import com.tinqinacademy.hotel.core.processors.BaseOperationProcessor;
 import com.tinqinacademy.hotel.persistence.model.Booking;
 import com.tinqinacademy.hotel.persistence.model.Guest;
@@ -16,7 +16,6 @@ import com.tinqinacademy.hotel.persistence.repository.GuestRepository;
 import io.vavr.control.Either;
 import io.vavr.control.Try;
 import jakarta.validation.Validator;
-import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.core.convert.ConversionService;
 import org.springframework.stereotype.Service;
